refactor(convert): extract convert
