+++ index.js
@@ -1,13 +1,12 @@
 'use strict';
 var request = require('request');
 var Promise = Promise || require('es6-promise').Promise;
-var cheerio = require('cheerio');
 var objectAssign = require('object-assign');
 var isOk = require('is-ok');
 
-var errors = require('./src/error');
-var HTTPError = errors.HTTPError;
-var InvalidFormatError = errors.InvalidFormatError;
+var HTTPError = require('./src/error').HTTPError;
+var InvalidFormatError = require('./src/error').InvalidFormatError;
+var convert = require('./src/convert');
 
 var defaultConfig = {
   headers: {
@@ -60,35 +59,6 @@ function mergeConfig(url, userConfig) {
 
 function buildDetailUrl(identifier) {
   return 'https://chrome.google.com/webstore/detail/' + identifier;
-}
-
-function convert(detailHtml) {
-  return new Promise(function (resolve, reject) {
-    var $ = cheerio.load(detailHtml);
-    var itemProps = {};
-    $('meta[itemprop]').each(function (index, element) {
-      // Split content like <meta itemprop="interactionCount" content="UserDownloads:418" />
-      if ($(element).attr('itemprop') === 'interactionCount' &&
-        $(element).attr('content').indexOf(':') !== -1) {
-        var keyValue = $(element).attr('content').split(':', 2);
-        itemProps[$(element).attr('itemprop')] = itemProps[$(element).attr('itemprop')] || {};
-        itemProps[$(element).attr('itemprop')][keyValue[0]] = keyValue[1];
-      } else {
-        itemProps[$(element).attr('itemprop')] = $(element).attr('content');
-      }
-    });
-    if (Object.keys(itemProps).length === 0) {
-      reject(new InvalidFormatError('There is no meta property'));
-      return;
-    }
-    if (!itemProps.hasOwnProperty('url') || !itemProps['url']) {
-      reject(new InvalidFormatError('url in response is required'));
-      return;
-    }
-    var splitUrl = itemProps.url.split('/');
-    itemProps['id'] = splitUrl[splitUrl.length - 1];
-    resolve(itemProps);
-  });
 }
 
 module.exports = run;
